nfl/fantasy/matchup/pyproject.toml
[project]
name = "matchup"
version = "0.1.0"
description = ""
authors = [
    {name = "Lukas Nesheim",email = "lukenesh@me.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "supabase (>=2.17.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "matchup", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
